+++ fltk/util/scaling/scaler.py
